Make setup.py not executable
Remove executable bit and remove shebang line.  Completely subjective
matter, and a matter of taste, but it doesn't make too much sense to
have the install script be executable, and it offers more flexibility
and clarity to invoke it with the interpreter explicitly.  You won't put
it on your PATH and call it all the time.